REFACTOR: 자동완성기능파일만 우선적으로 mvc패턴으로 변경 완료
