+++ app.js
@@ -8,7 +8,6 @@ app.use(express.static(__dirname));
 app.use("/fakeDB", fakeDB_router);
 
 app.get("/", (req, res) => {
-  // res.sendFile(__dirname + "/index.html");
   res.sendFile(path.join(__dirname, "index.html"));
 });
